user_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
passlib==1.7.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
httpx==0.25.2